DBZ-1723 Moved task logic of postgres and sqlserver (EventSourceBased) to facades objects to decouple and allow maintaining its state from outside of connect's object. This allows to retry in case of error with full state restart which is neccessary as db connections are spread accros multiple objects and needs to be reinitialized in case of db connection crash.
