+++ debezium-connector-postgres/src/main/java/io/debezium/connector/postgresql/PostgresConnectorTask.java
@@ -6,273 +6,24 @@
 
 package io.debezium.connector.postgresql;
 
-import java.nio.charset.Charset;
-import java.sql.SQLException;
-import java.time.Duration;
-import java.util.List;
-import java.util.Map;
-import java.util.concurrent.atomic.AtomicReference;
-import java.util.stream.Collectors;
-
-import org.apache.kafka.connect.errors.ConnectException;
-import org.apache.kafka.connect.source.SourceRecord;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 import io.debezium.config.Configuration;
 import io.debezium.config.Field;
-import io.debezium.connector.base.ChangeEventQueue;
-import io.debezium.connector.common.BaseSourceTask;
-import io.debezium.connector.postgresql.connection.PostgresConnection;
-import io.debezium.connector.postgresql.connection.ReplicationConnection;
-import io.debezium.connector.postgresql.spi.SlotCreationResult;
-import io.debezium.connector.postgresql.spi.SlotState;
-import io.debezium.connector.postgresql.spi.Snapshotter;
-import io.debezium.pipeline.ChangeEventSourceCoordinator;
-import io.debezium.pipeline.DataChangeEvent;
-import io.debezium.pipeline.ErrorHandler;
-import io.debezium.pipeline.EventDispatcher;
-import io.debezium.relational.TableId;
-import io.debezium.schema.TopicSelector;
-import io.debezium.util.Clock;
-import io.debezium.util.LoggingContext;
-import io.debezium.util.Metronome;
+import io.debezium.connector.common.ChangeEventSourceBasedTask;
+import io.debezium.pipeline.ChangeEventSourceFacade;
 
 /**
  * Kafka connect source task which uses Postgres logical decoding over a streaming replication connection to process DB changes.
  *
  * @author Horia Chiorean (hchiorea@redhat.com)
  */
-public class PostgresConnectorTask extends BaseSourceTask {
+public class PostgresConnectorTask extends ChangeEventSourceBasedTask {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(PostgresConnectorTask.class);
-    private static final String CONTEXT_NAME = "postgres-connector-task";
-
-    private static enum State {
-        RUNNING,
-        STOPPED;
-    }
-
-    private final AtomicReference<State> state = new AtomicReference<State>(State.STOPPED);
-
-    private volatile PostgresTaskContext taskContext;
-    private volatile ChangeEventQueue<DataChangeEvent> queue;
-    private volatile PostgresConnection jdbcConnection;
-    private volatile ChangeEventSourceCoordinator coordinator;
-    private volatile ErrorHandler errorHandler;
-    private volatile PostgresSchema schema;
-    private volatile Map<String, ?> lastOffset;
+    private PostgresChangeEventSourceFacade postgresChangeEventSourceFacade;
 
     @Override
-    public void start(Configuration config) {
-        if (!state.compareAndSet(State.STOPPED, State.RUNNING)) {
-            LOGGER.info("Connector has already been started");
-            return;
-        }
-
-        final PostgresConnectorConfig connectorConfig = new PostgresConnectorConfig(config);
-        final TopicSelector<TableId> topicSelector = PostgresTopicSelector.create(connectorConfig);
-        final Snapshotter snapshotter = connectorConfig.getSnapshotter();
-
-        if (snapshotter == null) {
-            throw new ConnectException("Unable to load snapshotter, if using custom snapshot mode, double check your settings");
-        }
-
-        jdbcConnection = new PostgresConnection(connectorConfig.jdbcConfig());
-        final TypeRegistry typeRegistry = jdbcConnection.getTypeRegistry();
-        final Charset databaseCharset = jdbcConnection.getDatabaseCharset();
-
-        schema = new PostgresSchema(connectorConfig, typeRegistry, databaseCharset, topicSelector);
-        this.taskContext = new PostgresTaskContext(connectorConfig, schema, topicSelector);
-        final PostgresOffsetContext previousOffset = (PostgresOffsetContext) getPreviousOffset(new PostgresOffsetContext.Loader(connectorConfig));
-        final Clock clock = Clock.system();
-
-        final SourceInfo sourceInfo = new SourceInfo(connectorConfig);
-        LoggingContext.PreviousContext previousContext = taskContext.configureLoggingContext(CONTEXT_NAME);
-        try {
-            // Print out the server information
-            SlotState slotInfo = null;
-            try (PostgresConnection connection = taskContext.createConnection()) {
-                if (LOGGER.isInfoEnabled()) {
-                    LOGGER.info(connection.serverInfo().toString());
-                }
-                slotInfo = connection.getReplicationSlotState(connectorConfig.slotName(), connectorConfig.plugin().getPostgresPluginName());
-            }
-            catch (SQLException e) {
-                LOGGER.warn("unable to load info of replication slot, Debezium will try to create the slot");
-            }
-
-            if (previousOffset == null) {
-                LOGGER.info("No previous offset found");
-                // if we have no initial offset, indicate that to Snapshotter by passing null
-                snapshotter.init(connectorConfig, null, slotInfo);
-            }
-            else {
-                LOGGER.info("Found previous offset {}", sourceInfo);
-                snapshotter.init(connectorConfig, previousOffset.asOffsetState(), slotInfo);
-            }
-
-            ReplicationConnection replicationConnection = null;
-            SlotCreationResult slotCreatedInfo = null;
-            if (snapshotter.shouldStream()) {
-                boolean shouldExport = snapshotter.exportSnapshot();
-                replicationConnection = createReplicationConnection(this.taskContext, shouldExport,
-                        connectorConfig.maxRetries(), connectorConfig.retryDelay());
-
-                // we need to create the slot before we start streaming if it doesn't exist
-                // otherwise we can't stream back changes happening while the snapshot is taking place
-                if (slotInfo == null) {
-                    try {
-                        slotCreatedInfo = replicationConnection.createReplicationSlot().orElse(null);
-                    }
-                    catch (SQLException ex) {
-                        throw new ConnectException(ex);
-                    }
-                }
-                else {
-                    slotCreatedInfo = null;
-                }
-            }
-
-            queue = new ChangeEventQueue.Builder<DataChangeEvent>()
-                    .pollInterval(connectorConfig.getPollInterval())
-                    .maxBatchSize(connectorConfig.getMaxBatchSize())
-                    .maxQueueSize(connectorConfig.getMaxQueueSize())
-                    .loggingContextSupplier(() -> taskContext.configureLoggingContext(CONTEXT_NAME))
-                    .build();
-
-            errorHandler = new ErrorHandler(PostgresConnector.class, connectorConfig.getLogicalName(), queue, this::cleanupResources);
-
-            final PostgresEventMetadataProvider metadataProvider = new PostgresEventMetadataProvider();
-
-            final EventDispatcher<TableId> dispatcher = new EventDispatcher<>(
-                    connectorConfig,
-                    topicSelector,
-                    schema,
-                    queue,
-                    connectorConfig.getTableFilters().dataCollectionFilter(),
-                    DataChangeEvent::new,
-                    PostgresChangeRecordEmitter::updateSchema,
-                    metadataProvider);
-
-            coordinator = new ChangeEventSourceCoordinator(
-                    previousOffset,
-                    errorHandler,
-                    PostgresConnector.class,
-                    connectorConfig,
-                    new PostgresChangeEventSourceFactory(
-                            connectorConfig,
-                            snapshotter,
-                            jdbcConnection,
-                            errorHandler,
-                            dispatcher,
-                            clock,
-                            schema,
-                            taskContext,
-                            replicationConnection,
-                            slotCreatedInfo),
-                    dispatcher,
-                    schema);
-
-            coordinator.start(taskContext, this.queue, metadataProvider);
-        }
-        finally {
-            previousContext.restore();
-        }
-    }
-
-    public ReplicationConnection createReplicationConnection(PostgresTaskContext taskContext, boolean shouldExport,
-                                                             int maxRetries, Duration retryDelay)
-            throws ConnectException {
-        final Metronome metronome = Metronome.parker(retryDelay, Clock.SYSTEM);
-        short retryCount = 0;
-        ReplicationConnection replicationConnection = null;
-        while (retryCount <= maxRetries) {
-            try {
-                return taskContext.createReplicationConnection(shouldExport);
-            }
-            catch (SQLException ex) {
-                retryCount++;
-                if (retryCount > maxRetries) {
-                    LOGGER.error("Too many errors connecting to server. All {} retries failed.", maxRetries);
-                    throw new ConnectException(ex);
-                }
-
-                LOGGER.warn("Error connecting to server; will attempt retry {} of {} after {} " +
-                        "seconds. Exception message: {}", retryCount, maxRetries, retryDelay.getSeconds(), ex.getMessage());
-                try {
-                    metronome.pause();
-                }
-                catch (InterruptedException e) {
-                    LOGGER.warn("Connection retry sleep interrupted by exception: " + e);
-                    Thread.currentThread().interrupt();
-                }
-            }
-        }
-        return replicationConnection;
-    }
-
-    @Override
-    public void commit() throws InterruptedException {
-        if (coordinator != null) {
-            coordinator.commitOffset(lastOffset);
-        }
-    }
-
-    @Override
-    protected List<SourceRecord> pollRecords() throws InterruptedException {
-        final List<DataChangeEvent> records = queue.poll();
-
-        final List<SourceRecord> sourceRecords = records.stream()
-                .map(DataChangeEvent::getRecord)
-                .collect(Collectors.toList());
-
-        if (!sourceRecords.isEmpty()) {
-            this.lastOffset = sourceRecords.get(sourceRecords.size() - 1).sourceOffset();
-        }
-
-        return sourceRecords;
-    }
-
-    @Override
-    public void stop() {
-        cleanupResources();
-    }
-
-    private void cleanupResources() {
-        if (!state.compareAndSet(State.RUNNING, State.STOPPED)) {
-            LOGGER.info("Connector has already been stopped");
-            return;
-        }
-
-        try {
-            if (coordinator != null) {
-                coordinator.stop();
-            }
-        }
-        catch (InterruptedException e) {
-            Thread.currentThread().interrupt();
-            LOGGER.error("Interrupted while stopping coordinator", e);
-            throw new ConnectException("Interrupted while stopping coordinator, failing the task");
-        }
-
-        try {
-            if (errorHandler != null) {
-                errorHandler.stop();
-            }
-        }
-        catch (InterruptedException e) {
-            Thread.currentThread().interrupt();
-            LOGGER.error("Interrupted while stopping", e);
-        }
-
-        if (jdbcConnection != null) {
-            jdbcConnection.close();
-        }
-
-        if (schema != null) {
-            schema.close();
-        }
+    protected ChangeEventSourceFacade createEventSourceFacade(Configuration config) {
+        postgresChangeEventSourceFacade = new PostgresChangeEventSourceFacade(config, loader -> getPreviousOffset(loader));
+        return postgresChangeEventSourceFacade;
     }
 
     @Override
@@ -285,7 +36,11 @@ public class PostgresConnectorTask extends BaseSourceTask {
         return PostgresConnectorConfig.ALL_FIELDS;
     }
 
-    public PostgresTaskContext getTaskContext() {
-        return taskContext;
+    /**
+     * TODO remove this used only in test
+     */
+    PostgresTaskContext getTaskContext() {
+        return postgresChangeEventSourceFacade.getTaskContext();
     }
+
 }
