+++ debezium-core/src/main/java/io/debezium/pipeline/ChangeEventSourceCoordinator.java
@@ -109,7 +109,7 @@ public class ChangeEventSourceCoordinator {
                 LOGGER.warn("Change event source executor was interrupted", e);
             }
             catch (Throwable e) {
-                errorHandler.setProducerThrowable(e);
+                errorHandler.handleThrowable(e);
             }
             finally {
                 streamingMetrics.connected(false);
